test: share support and add orchestration coverage
