+++ src/executor.rs
@@ -933,10 +933,7 @@ async fn execute_single_step(
 
 #[cfg(test)]
 mod tests {
-    #[allow(dead_code)]
-    mod common {
-        include!(concat!(env!("CARGO_MANIFEST_DIR"), "/tests/common/mod.rs"));
-    }
+    use crate::test_support as common;
 
     use super::*;
     use crate::provider::{OutputChunk, ProviderError};
