+++ src/lib.rs
@@ -16,4 +16,5 @@ pub mod provider;
 pub mod report;
 pub mod run;
 pub mod test_file;
+pub mod test_support;
 pub mod update;
